// apps/web/src/lib/api/generated/model/schemeOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BrokerOS API
 * OpenAPI spec version: 1.0.0
 */
import type { ResetPeriod } from './resetPeriod';

export interface SchemeOut {
  branch_id: string | null;
  document_type: string;
  id: string;
  is_active: boolean;
  pattern: string;
  reset_period: ResetPeriod;
  start_at: number;
  version: number;
}